# [feat](clone) Speed clone tablet via batch small file downloading