--- a/fe/fe-core/src/main/java/org/apache/doris/persist/BatchDropInfo.java
+++ b/fe/fe-core/src/main/java/org/apache/doris/persist/BatchDropInfo.java
@@ -26,6 +26,7 @@ import com.google.gson.annotations.SerializedName;
 import java.io.DataInput;
 import java.io.DataOutput;
 import java.io.IOException;
+import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
 
@@ -43,12 +44,15 @@ public class BatchDropInfo implements Writable {
     private String tableName; // not used in equals and hashCode
     @SerializedName(value = "indexIdSet")
     private Set<Long> indexIdSet;
+    @SerializedName(value = "indexNameMap")
+    private Map<Long, String> indexNameMap; // not used in equals and hashCode
 
-    public BatchDropInfo(long dbId, long tableId, String tableName, Set<Long> indexIdSet) {
+    public BatchDropInfo(long dbId, long tableId, String tableName, Map<Long, String> indexNameMap) {
         this.dbId = dbId;
         this.tableId = tableId;
         this.tableName = tableName;
-        this.indexIdSet = indexIdSet;
+        this.indexIdSet = indexNameMap.keySet();
+        this.indexNameMap = indexNameMap;
     }
 
     @Override
@@ -80,6 +84,10 @@ public class BatchDropInfo implements Writable {
 
     public Set<Long> getIndexIdSet() {
         return indexIdSet;
+    }
+
+    public Map<Long, String> getIndexNameMap() {
+        return indexNameMap;
     }
 
     public long getDbId() {